Preserve `Subst` literals
If we propagate the same literal with multiple origins in a single CC(X)
step (i.e. they end up in the same fact list), we only keep one. We must
make sure to keep the one with `Subst` origin, because those are used by
the union-find communicate with the relations about representative
changes.

Fixes #883