Using `dune build @check` to build the project (#887)
The current target `all` of the `Makefile` is a bit complicated as
we want to build only the codebase without running tests when we type
`make`. After looking for a better solution, I found this issue on the
`dune` GitHub (https://github.com/ocaml/dune/issues/7203) explaining
we have to use `dune build @check` and `dune build` builds everything
by design.

I modify the `Makefile` accordingly.