Locking the Dolmen to a dev commit
This commit locks the Dolmen used by AE to a specific commit
in order to use the unreleased feature for custom statements.

I also notice that the `with_cache` in `d_cnf` was silently shadow by
the same function from the module `Expr` of `Dolmen`. The signature of
this latter changed recently. I fix it by using the `with_cache` from
Dolmen. Notice that the hash table of the cache is hidden in the closure
of `with_cache` and `bv2nat` and `int2bv` don't share their cache as
before.