Fix the target `all` of the Makefile (#914)
The PR #887 was wrong... The target @check of `Dune` doesn't build
all the binaries, it builds only binary artefacts `cmi`, `cmx`, ... but
no linkage is performed.

This commit mostly reverts the changes. As the sources are located in
`src/`, I use the target `@src/all`. Now the command `make` will build
everything in the directory `src`. I also add `examples` since we want
to check the libusage module still compile.