Adding a test for optimization in SatML frontend
The current implementation of the optimization feature is
a bit complicated. In particular, the feature requires that the
SAT solver keeps exploring other SAT branches of the problem to
be sure to get an optimized model.

This commit adds a nonregression test for the support of optimization in
the frontend of SatML.