Remove deadcode optimization in `look_for_sat`
The `look_for_sat` contains a commented case to process the input choice
that isn't valid because the actual explanation doesn't contain the
appropriate information. As we plan to refactor completely this part of
the code, it doesn't make sense to keep it anymore.

This commit remove completely this optimization and the flag `bad_case`
used to activate it.