Simplify `mk_const_term`
Removing the conversion to `string` of the input rational `c` in the
function `mk_const_term`. The smart constructors `Expr.Reals.of_Q` and
`Expr.Ints.of_Z` still do a round-trip through the type `string` but
this commit reduces the number of locations in the codebase where we do
this conversion.