Fix issue 926
This PR fixes the issue #926 about printing reals in models.

The current printer always displays positive reals as rational numbers of the
form `(/ x y)` where `x` and `y` are positive integers. Negative reals are
represented by the expression `(/ (- x) y)` where `x` and `y` are positive
integers. The printer simplifies the expression if the denominator is one.

Notice that by the current implementation of Zarith:
```ocaml
type t = {
    num: Z.t; (** Numerator. *)
    den: Z.t; (** Denominator, >= 0 *)
  }
(* Type of rationals.
   Invariants:
   - den is always >= 0;
   - num and den have no common factor;
   - if den=0, then num is -1, 0 or 1.
   - if num=0, then den is -1, 0 or 1.
 *)
```
Any rational number of the form `0/q` satisfies `q = -1` or `q = 0` or `q = 1`.
Thus the printer will always output `0.0` in this case but for the case `0/0`
which shouldn't occur.