Do not test .smt2 files with the legacy frontend
Since we don't officially support the SMT-LIB input with the legacy
frontend, and that the legacy frontend is deprecated and scheduled for
removal, there really is no need to test SMT-LIB input with the legacy
frontend.

This means we no longer will have to tag `.smt2` tests with `.dolmen.`
just because we use some SMT-LIB features (such as bit-vector
primitives) that are not supported by the legacy frontend.