Always allow polymorphic input in smt-lib format
This makes the Dolmen frontend behave like the legacy frontend and
always allows polymorphic extensions of smt-lib input (this means .smt2
and .psmt2 files as well as the `--input smtlib2` option).

As pointed out by @Gbury this means that we need to override Dolmen's
language inference for `.smt2` files. We also need to properly take care
of `.zip` input.

Fixes #933