feat: Allow interactive SMT-LIB input
Dolmen supports interactive SMT-LIB input on stdin, but we currently do
not make use of it because we *first* read everything from stdin
(awaiting a ^D end-of-file from the user), then pass that to Dolmen,
making things useless for interactive input.

This patch makes it so that interactive input works properly (provided
you also set `--continue-on-error`, and maybe redirect `stderr` to
oblivion).

Note that Dolmen does not support reading from `stdin` in the native
format, so we keep the old mode of operations if the user explicitly
requested `--input native`. Otherwise, we default to SMT-LIB format when
reading from stdin, and set the output format to SMT-LIB by default.

Changing the default format for stdin is not a breaking change because
previously we were rejecting stdin input without an explicit format with
an error:

> Error The following extension was not recognized: ''.
>      Please use a recognised extension or specify an input language \
on the command line