Bubble printing model (#932)
* Bubbling the model printing in the binary

This commit bubbles the decision of printing models in the binary. Now,
the library do not print models at all.

* Update the JS worker

We have to update the JS worker to dump models as we don't print models
in the library anymore.

* Save a timeout reason in the exception Util.Timeout

In some situations, we need to retrieve the reason of a timeout in the
binary but we have no environment of the SAT solver in these locations.

I add an argument to the timeout exception to save its reason.

* Revert "Save a timeout reason in the exception Util.Timeout"

This reverts commit dbfed1b65a7636a359e7aee22d03713018529070.

* Review changes

* Factorizing `print_model`

We use `print_model` twice (in the usual binary and the JS worker), thus
we can factorize it by moving it in `Frontend`.