fix(BV): Fix unsoundness in BV model generation
The BV model generation was using an incorrect understanding of the
[is_cs] flag that lead to unsoundness because it assumed that it was
fine to set this flag to [false] when the value is forced *in the
current context* whereas it is only valid to do so when the value is
forced *in all contexts*.

This PR fixes that, and adds copious amounts of documentation to explain
the [is_cs] flag so that we don't make the same mistake again.

Unfortunately the fix leads to a test that no longer passes, but for
good reason as the reasoning used for it was unsound. Hopefully this
should be fixed in a follow-up to
https://github.com/OCamlPro/alt-ergo/pull/944

I have also added a new test that is actually satisfiable but would
previously return [unsat]. It times out for now, but is solved properly
in https://github.com/OCamlPro/alt-ergo/pull/944