fix(BV): Do not call [X.make] on non-subterm, reloaded
This is a follow-up to https://github.com/OCamlPro/alt-ergo/pull/954

Somehow convinced myself that the cases where we perform simplifications
were fine because we would only be simplifying into a subterm, but that
is obviously not the case due to negation. Let's just do the safe thing
and always return the original term with the appropriate equalities.

All of this code should be removed in favor of doing these computations
in the relations instead anyways, which would make sure we always treat
[bv2nat] in the same way for literals and after substitution (see also
https://github.com/OCamlPro/alt-ergo/issues/824 )