Handle case-split conflicts properly
We cannot simply claim that forced splits are not case splits (see
https://github.com/OCamlPro/alt-ergo/pull/968 ). In order to do proper
backtracking using exlpanations, we have to propagate the equality
[x = #b1] when we learn [x = #b0] from the case-split.

To keep things simple, we only directly support [Distinct] in the BV
domain for the literals that come from a case split (other cases are
handled by the union-find *after* the splits). We might want to add
proper support for arbitrary [Distinct] constraints later.