Simplify smart constructors for literal numbers
We made a lot of conversion between string and Zarith representation for
literals numbers in `Expr`. This commit simplifies the situation.

Now we perform the conversion of the string representation into Zarith
one once in both legacy and Dolmen frontend.

The motivation for this modification comes from the fact we'll use
expressions to store model values. If we use string here, it's not
easy to normalize terms like `-(- 0 x)` to `x`.

With the Zarith representation, the library does the normalization for
us.

Remove the smart constructors `int` and `real` in `Expr` which
aren't useful anymore.