Replace `Symbols.t` by `Var.t` in binders
Alt-Ergo embeds variables into symbols then expressions. It's a good
design but at many places in the codebase, we use symboles where
variables are clearly sufficient.

For instance, we shouldn't use symbols in substitutions or binders
because we always bind a variable to an expression or substitute a
variable by an expression.

Let's use the simpler type `Var.t` as much as possible.

This commit should be safe because the comparison functions of
`Symbols.t` and `Var.t` agree on variables.