fix(BV): Allow substituting constants (#979)
Substituting a constant should be a no-op (just like substituting a
variable that is not in the substitution). Somehow this case was missed
in the new version of `apply_sub`. It is fairly hard to trigger this
case however: it can only happen in `apply_subs`, and requires a
C-substitution to be created, which in turn requires a C-variable to be
split during B-variable slicing, and then the C-substitution must be
applied to an equation rooted in a distinct term that contains a
constant. This patch includes a test case where this happens.