Use a fixed name for Var.underscore
There is no good reason to use a random number in the name of the
`underscore` variable; this can cause reproducibility issues and
potentially accidental unsoundness (I have not looked into details into
how `underscore` is used, possibly it would lead to crashes rather than
unsoundness, but both are bad anyways).

This patch always uses `._` instead, which is a solver-reserved name.