cleanup: Use proper namespacing for names
This introduces proper namespacing for variable names instead of relying
on string prefixes. In order to disturb the ordering as little as
possible, names are pre-mangled using their namespace (e.g. "x" in the
`Internal` namespace is stored as ".!x") and further operations
(hashing, comparison, etc.) are performed on the mangled names.