cleanup: Use proper ADT for variable classes (#986)
This patch cleans up the way we use the '?' (question mark) and '_'
special variables by introducing special constructors in
`Symbols.var_bnd` and `Var.t` for their corresponding purposes.

Note that it looks like this fixes a subtle bug in `IntervalCalculus`,
which assumes that all variables whose name starts with a '?' are
"local" variables from semantic triggers, while that is not true when
using the SMT-LIB format. However, since that piece of code only runs
during semantic matching, it would require defining theory extensions in
the SMT-LIB format to trigger the bug, and since that is not supported,
there is in fact no bug.