cleanup(CDCL): Use a set of flat formulas to activate
The `activate` field in the `Satml_frontend` is currently a map from
flat formulas to `Atom.atom option`, but the `Atom.atom option` is
always `None`, and having anything else than `None` in that map would
result in assertion failures.

This patch replaces that map by a set, since we no longer know what the
intended meaning for the map was.