refactor(CDCL): Make the Heap module polymorphic
This patch refactors the implementation of the `Iheap` module to allow
arbitrary elements rather than integers. This allows dropping the
requirement to index the heap by variable identifiers in the CDCL
solver, making the `vars` vector always dense.

The implementation is intrusive: each variable can only be in a single
heap at once.