cleanup(Vec): Make vectors dense
The only reason for vectors to forbid dummy elements in the reachable
part of the vector was to accomodate the case of the variables heap,
which could have holes.

Since the heap can no longer have holes, this functionality is no longer
needed, and it is no longer necessary to leak the existence of the dummy
element to the user (outside of vector-constructing functions).