Make the `Heap` module polymorphic (#1006)
This patch refactors the implementation of the `Iheap` module to allow
arbitrary elements rather than integers. This allows dropping the
requirement to index the heap by variable identifiers in the CDCL
solver, making the `vars` vector always dense.

The implementation is intrusive: each variable can only be in a single
heap at once.

The only reason for vectors to forbid dummy elements in the reachable
part of the vector was to accomodate the case of the variables heap,
which could have holes.

Since the heap can no longer have holes, this functionality is no longer
needed, and it is no longer necessary to leak the existence of the dummy
element to the user (outside of vector-constructing functions).