doc(CDCL): Document the CDCL-Tableaux solver (#995)
This is a documentation-only patch for the CDCL-Tableaux solver. There
is not much additional documentation for the pure CDCL solver, except
regarding the use of the `tatoms_queue` as an intermediate place to hold
decided and (clause) propagated atoms before performing theory
propagations in bulk. The CDCL solver is heavily based on miniSAT, and
hence http://www.decision-procedures.org/handouts/MiniSat.pdf is
probably a better documentation at a high level than anything we could
write.