fix(BV): Actually propagate freshness change
The `Constraints` module has a mechanism to remember fresh constraints
that have never been propagated, so that we don't need to always
re-propagate constraints that have already been propagated if their
arguments have not changed. Unfortunately, we forget the change to this
set of fresh constraints, and hence always re-propagate all active
constraints, which renders the optimization useless.

This patch actually takes into consideration the changes to the set of
fresh constraints.